refactor(report-page): rename folder datasets-report-page to report-page
