+++ src/pages/report-page/publishers-select/publishers-select.scss
