+++ src/pages/report-page/publishers-tree/publishers-tree.component.jsx
