+++ src/pages/report-page/publishers-tree/publishers-tree.scss
