+++ src/pages/report-page/report-page-resolver.js
