+++ src/pages/report-page/report-stats/chart-bar/chart-bar.scss
