+++ src/pages/report-page/report-stats/concept-stats/concept-stats.component.jsx
