+++ src/pages/report-page/report-stats/report-stats.scss
