+++ src/pages/report-page/report-stats/stat-box/stat-box.scss
